Cross-compile for Scala 2.10, 2.11 and 2.12
As most of the Scala users now use Scala 2.12, it only
makes sense to support it. For those stuck with older
versions, Scala 2.11 and 2.10 is also supported.

As Maven doesn't have a nice way to support cross-compilation
for Scala, I have done it in SBT, with keeping the maven build
around.